Login and Registration files
I did not have time to make sure there are no bugs, so please run this and let me know if something doesn't work properly. Everything should be fine, though. There are no toast messages for when you log in successfully, so we would probably need a small pop-up saying welcome or something like that. Then we do need a logout feature, which is just a pop-up asking if they want to log out and then just doing stuff and sending them back to the login screen. Otherwise, everything should be fine.